TH2-2150. Added limits to message and test event batches. These settings are required by CradleStorage. Entities that comform to this can be built with cradleStorage.getEntitiesFactory().

+++ cradle-core/src/main/java/com/exactpro/cradle/messages/MessageBatch.java
@@ -43,7 +43,7 @@ public interface MessageBatch
 	/**
 	 * @return size of messages currently stored in the batch
 	 */
-	long getBatchSize();
+	int getBatchSize();
 
 	/**
 	 * @return collection of messages stored in the batch
